+++ src/main/java/com/uj15/timedeal/common/auth/AuthenticationResolver.java
@@ -1,6 +1,7 @@
 package com.uj15.timedeal.common.auth;
 
 import com.uj15.timedeal.auth.SessionConst;
+import com.uj15.timedeal.common.auth.annotation.Authentication;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.core.MethodParameter;
