+++ src/main/java/com/uj15/timedeal/common/auth/annotation/Authorization.java
@@ -1,4 +1,4 @@
-package com.uj15.timedeal.common.auth;
+package com.uj15.timedeal.common.auth.annotation;
 
 import com.uj15.timedeal.user.Role;
 import java.lang.annotation.ElementType;
